#993. Use unsigned types to avoid little/big endian problems

+++ LibTest/ffi/Char/Char_A01_t02.dart
@@ -23,7 +23,7 @@ import "../../../Utils/expect.dart";
 
 void main() {
   Pointer<Char> p1 = calloc<Char>();
-  Pointer<Int16> p2 = new Pointer<Int16>.fromAddress(p1.address);
+  Pointer<Uint8> p2 = new Pointer<Uint8>.fromAddress(p1.address);
   try {
     p2.value = 0;
     Expect.equals(0, p1.value);
